Ajuste permissões e visualizaçao de telas p socios e funcionarios

+++ app/Http/Controllers/FuncionarioController.php
@@ -10,6 +10,10 @@ class FuncionarioController extends Controller
     // Exibe uma lista de todos os funcionários
     public function index(Request $request)
     {
+        if (auth()->user()->perfil_acesso !== 'socio') {
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página.');
+        }
+
         $query = Funcionario::query();
     
         // Aplicando filtro por nome, cargo e perfil de acesso
@@ -34,12 +38,22 @@ class FuncionarioController extends Controller
     // Exibe o formulário de criação de funcionários
     public function create()
     {
+
+        if (auth()->user()->perfil_acesso !== 'socio') {
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página.');
+        }
+
         return view('cadastroFuncionario');
     }
 
     // Armazena um novo funcionário
     public function store(Request $request)
     {
+
+        if (auth()->user()->perfil_acesso !== 'socio') {
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página.');
+        }
+
         $funcionario = new Funcionario();
 
         $funcionario->nome = $request->input('nome');
@@ -50,12 +64,16 @@ class FuncionarioController extends Controller
 
         $funcionario->save();
 
-        return redirect('/funcionarios');
+        return redirect('/funcionarios')->with('msg', 'Funcionário criado com sucesso!');
     }
 
     // Exibe o formulário de edição de um funcionário específico
     public function edit($id)
     {
+        if (auth()->user()->perfil_acesso !== 'socio') {
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página.');
+        }
+
         $funcionario = Funcionario::findOrFail($id);
         return view('editaFuncionario', compact('funcionario'));
     }
@@ -63,6 +81,10 @@ class FuncionarioController extends Controller
     // Atualiza um funcionário específico
     public function update(Request $request, $id)
     {
+        if (auth()->user()->perfil_acesso !== 'socio') {
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página.');
+        }
+
         $funcionario = Funcionario::findOrFail($id);
 
         $funcionario->nome = $request->input('nome');
@@ -75,15 +97,20 @@ class FuncionarioController extends Controller
 
         $funcionario->save();
 
-        return redirect('/funcionarios');
+        return redirect('/funcionarios')->with('msg', 'Funcionário atualizado com sucesso!');
     }
 
     // Exclui um funcionário específico
     public function destroy($id)
     {
+
+        if (auth()->user()->perfil_acesso !== 'socio') {
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página.');
+        }
+        
         $funcionario = Funcionario::findOrFail($id);
         $funcionario->delete();
         
-        return redirect('/funcionarios');
+        return redirect('/funcionarios')->with('msg', 'Funcionário excluído com sucesso!');
     }
 }
